# server/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Luminaris API",
    "version": "1.0.0",
    "description": "Document Intelligence Platform API"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Invalid request payload or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/accounting/{unitId}/periods": {
      "get": {
        "summary": "List accounting periods for a fiscal year",
        "parameters": [
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounting periods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/accounting/{unitId}/periods/seed-year": {
      "post": {
        "summary": "Seed 12 FUTURE periods for a fiscal year",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedYearInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Periods seeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/accounting/periods/{id}/open": {
      "post": {
        "summary": "Open a FUTURE or SOFT_CLOSED period → OPEN",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Period id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "unitId"
                ],
                "properties": {
                  "unitId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period opened"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "description": "periodSemantics: as_of. Transition recorded in AccountingPeriodTransition.",
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/accounting/periods/{id}/soft-close": {
      "post": {
        "summary": "Soft-close an OPEN period (can be reopened)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePeriodInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period soft-closed"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/accounting/periods/{id}/hard-close": {
      "post": {
        "summary": "Permanently close a period (HARD_CLOSED = terminal)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePeriodInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period hard-closed"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/accounting/periods/{id}/reopen": {
      "post": {
        "summary": "Reopen a SOFT_CLOSED period → OPEN (HARD_CLOSED is terminal)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenPeriodInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period reopened"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "tags": [
          "Accounting"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}